refactor: Convert types.KindInstaller to Handler format
